+++ wox.core/util/overlay/overlay_darwin.go
@@ -23,6 +23,10 @@ typedef struct {
     float height;
     float fontSize;
     float iconSize;
+    char* tooltip;
+    unsigned char* tooltipIconData;
+    int tooltipIconLen;
+    float tooltipIconSize;
 } OverlayOptions;
 
 void ShowOverlay(OverlayOptions opts);
@@ -67,6 +71,9 @@ func Show(opts OverlayOptions) {
 		cMessage := C.CString(opts.Message)
 		defer C.free(unsafe.Pointer(cMessage))
 
+		cTooltip := C.CString(opts.Tooltip)
+		defer C.free(unsafe.Pointer(cTooltip))
+
 		var cIconData *C.uchar
 		var cIconLen C.int
 
@@ -74,6 +81,14 @@ func Show(opts OverlayOptions) {
 		if len(pngBytes) > 0 {
 			cIconData = (*C.uchar)(unsafe.Pointer(&pngBytes[0]))
 			cIconLen = C.int(len(pngBytes))
+		}
+
+		var cTooltipIconData *C.uchar
+		var cTooltipIconLen C.int
+		tooltipPngBytes, _ := imageToPNG(opts.TooltipIcon)
+		if len(tooltipPngBytes) > 0 {
+			cTooltipIconData = (*C.uchar)(unsafe.Pointer(&tooltipPngBytes[0]))
+			cTooltipIconLen = C.int(len(tooltipPngBytes))
 		}
 
 		cOpts := C.OverlayOptions{
@@ -93,6 +108,10 @@ func Show(opts OverlayOptions) {
 			height:           C.float(opts.Height),
 			fontSize:         C.float(opts.FontSize),
 			iconSize:         C.float(opts.IconSize),
+			tooltip:          cTooltip,
+			tooltipIconData:  cTooltipIconData,
+			tooltipIconLen:   cTooltipIconLen,
+			tooltipIconSize:  C.float(opts.TooltipIconSize),
 		}
 
 		C.ShowOverlay(cOpts)