+++ CHANGELOG.md
@@ -4,12 +4,12 @@
 
 - Add
   - [`Calculator`] Add comma separator support in Calculator plugin #4325
-
+  - [`File Explorer Search`] Add query tips for file explorer plugin (Default is on, user can close this tip in plugin setting)
 - Improve
   - Improve everything sdk integration stability (with 1.5a support) #4317
 
 - Fix
-  - [`Current Folder`] Fix a issue that Current folder plugin's settings do not load #4326
+  - [`File Explorer Search`] Fix a issue that file explorer search plugin's settings do not load #4326
   - [`Clipboard`] Fix a issue that Clipboard plugin cannot paste to active window #4328
   - [`Wpm`] Fix a issue where WPM couldn't create script plugins #4330